Move commands to a separate view presenter
