+++ app/Presenters/ProjectPresenter.php
@@ -2,7 +2,6 @@
 
 namespace App\Presenters;
 
-use App\Command;
 use App\Project;
 use Lang;
 use Robbo\Presenter\Presenter;
@@ -10,7 +9,7 @@ use Robbo\Presenter\Presenter;
 /**
  * The view presenter for a project class.
  */
-class ProjectPresenter extends Presenter
+class ProjectPresenter extends CommandPresenter
 {
     /**
      * Returns the build status needed by CCTray
@@ -92,94 +91,6 @@ class ProjectPresenter extends Presenter
     }
 
     /**
-     * Gets the readable list of before clone commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentBeforeClone()
-    {
-        return $this->commandNames(Command::BEFORE_CLONE);
-    }
-
-    /**
-     * Gets the readable list of after clone commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentAfterClone()
-    {
-        return $this->commandNames(Command::AFTER_CLONE);
-    }
-
-    /**
-     * Gets the readable list of before install commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentBeforeInstall()
-    {
-        return $this->commandNames(Command::BEFORE_INSTALL);
-    }
-
-    /**
-     * Gets the readable list of after install commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentAfterInstall()
-    {
-        return $this->commandNames(Command::AFTER_INSTALL);
-    }
-
-    /**
-     * Gets the readable list of before activate commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentBeforeActivate()
-    {
-        return $this->commandNames(Command::BEFORE_ACTIVATE);
-    }
-
-    /**
-     * Gets the readable list of after activate commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentAfterActivate()
-    {
-        return $this->commandNames(Command::AFTER_ACTIVATE);
-    }
-
-    /**
-     * Gets the readable list of before purge commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentBeforePurge()
-    {
-        return $this->commandNames(Command::BEFORE_PURGE);
-    }
-
-    /**
-     * Gets the readable list of after purge commands.
-     *
-     * @return string
-     * @see self::commandNames()
-     */
-    public function presentAfterPurge()
-    {
-        return $this->commandNames(Command::AFTER_PURGE);
-    }
-
-    /**
      * Show the application status.
      *
      * @return string
@@ -231,6 +142,7 @@ class ProjectPresenter extends Presenter
 
     /**
      * The application heartbeat status css.
+     *
      * @return string
      */
     public function presentHeartBeatStatusCss()
@@ -244,28 +156,5 @@ class ProjectPresenter extends Presenter
         }
 
         return 'success';
-    }
-
-    /**
-     * Gets the readable list of commands.
-     *
-     * @param int $stage
-     * @return string
-     */
-    private function commandNames($stage)
-    {
-        $commands = [];
-
-        foreach ($this->object->commands as $command) {
-            if ($command->step === $stage) {
-                $commands[] = $command->name;
-            }
-        }
-
-        if (count($commands)) {
-            return implode(', ', $commands);
-        }
-
-        return Lang::get('app.none');
     }
 }
